use the frame pipeline in example 5

+++ src/renderer/clear_frame.rs
@@ -47,7 +47,7 @@ impl Renderer for ClearFrame {
         &self,
         vk_dev: &RenderDevice,
         cmd: vk::CommandBuffer,
-        current_image: u32,
+        current_image: usize,
     ) -> Result<()> {
         self.render_pass
             .begin_render_pass(vk_dev, cmd, current_image);
